Don't try to set any Redis config in production

+++ app.py
@@ -34,9 +34,17 @@ import urllib.parse
 import xml.etree.ElementTree as ET
 
 app = flask.Flask(__name__)
-cache = redis.from_url(os.environ.get("REDISCLOUD_URL", "redis://localhost"))
-cache.config_set("maxmemory-policy", "allkeys-lru")
-if app.debug: cache.flushdb()
+
+if "REDISCLOUD_URL" in os.environ:
+    # Production config values are set in the dashboard.
+    # See 'heroku addons:open rediscloud'.
+    cache = redis.from_url(os.environ["REDISCLOUD_URL"])
+else:
+    cache = redis.from_url("redis://localhost")
+    cache.config_set("maxmemory", "30mb")
+    cache.config_set("maxmemory-policy", "allkeys-lru")
+    if app.debug: cache.flushdb()
+
 
 @app.route("/favicon")
 def favicon():
